fix: Editor hook 분리
